+++ spec/dummy/app/controllers/dummy_controller.rb
@@ -1,6 +1,0 @@
-class DummyController < ApplicationController
-  include Agilibox::ApiControllerConcern
-
-  def show
-  end
-end